+++ Tests/WindowsCloudPC.Tests.ps1
@@ -4,7 +4,9 @@ BeforeDiscovery {
         'Get-CloudPC',
         'Get-CloudPCByProvisioningPolicy',
         'Get-CloudPCProvisioningPolicy',
-        'Get-CloudPCUsage'
+        'Get-CloudPCRemoteActionResult',
+        'Get-CloudPCUsage',
+        'Restart-CloudPC'
     )
 }